+++ test/common/seekablesubreadstream.h
@@ -20,7 +20,7 @@ class SeekableSubReadStreamTestSuite : public CxxTest::TestSuite
 		{
 			TS_ASSERT( !ssrs.eos() );
 
-			TS_ASSERT_EQUALS( i - start, ssrs.pos() );
+			TS_ASSERT_EQUALS( uint32(i - start), ssrs.pos() );
 
 			ssrs.read(&b, 1);
 			TS_ASSERT_EQUALS( i, b );
@@ -37,34 +37,34 @@ class SeekableSubReadStreamTestSuite : public CxxTest::TestSuite
 		Common::SeekableSubReadStream ssrs = Common::SeekableSubReadStream(&ms, 1, 9);
 		byte b;
 
-		TS_ASSERT_EQUALS( ssrs.pos(), 0 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)0 );
 
 		ssrs.seek(1, SEEK_SET);
-		TS_ASSERT_EQUALS( ssrs.pos(), 1 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)1 );
 		b = ssrs.readByte();
 		TS_ASSERT_EQUALS( b, 2 );
 
 		ssrs.seek(5, SEEK_CUR);
-		TS_ASSERT_EQUALS( ssrs.pos(), 7 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)7 );
 		b = ssrs.readByte();
 		TS_ASSERT_EQUALS( b, 8 );
 
 		ssrs.seek(-3, SEEK_CUR);
-		TS_ASSERT_EQUALS( ssrs.pos(), 5 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)5 );
 		b = ssrs.readByte();
 		TS_ASSERT_EQUALS( b, 6 );
 
 		ssrs.seek(0, SEEK_END);
-		TS_ASSERT_EQUALS( ssrs.pos(), 8 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)8 );
 		TS_ASSERT( ssrs.eos() );
 
 		ssrs.seek(3, SEEK_END);
-		TS_ASSERT_EQUALS( ssrs.pos(), 5 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)5 );
 		b = ssrs.readByte();
 		TS_ASSERT_EQUALS( b, 6 );
 
 		ssrs.seek(8, SEEK_END);
-		TS_ASSERT_EQUALS( ssrs.pos(), 0 );
+		TS_ASSERT_EQUALS( ssrs.pos(), (uint32)0 );
 		b = ssrs.readByte();
 		TS_ASSERT_EQUALS( b, 1 );
 	}
